requirements_server.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.1.8
django-celery-beat==2.4.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-timezone-field==5.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
exceptiongroup==1.0.0rc9
filelock==3.12.0
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
lap==0.4.0
lit==16.0.2
lxml==4.9.1
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-cupti-cu11==11.7.101
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
# nvidia-cufft-cu11==10.9.0.58
# nvidia-curand-cu11==10.2.10.91
# nvidia-cusolver-cu11==11.4.0.1
# nvidia-cusparse-cu11==11.7.4.91
# nvidia-nccl-cu11==2.14.3
# nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==21.3
pandas==2.0.1
phonenumbers==8.12.57
Pillow==9.3.0
prompt-toolkit==3.0.33
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
# scipy==1.10.1
# seaborn==0.12.2
# selenium==4.5.0
sentry-sdk==1.21.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
sympy==1.11.1
# tensorboard==2.12.2
# tensorboard-data-server==0.7.0
# tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
# trio==0.22.0
# trio-websocket==0.9.2
# triton==2.0.0
typing_extensions==4.5.0
tzdata==2022.6
ultralytics==8.0.48
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.3.0
whitenoise==6.2.0
wsproto==1.2.0
zipp==3.15.0
